Refactor calculation tests
